refactored time clock related code

+++ app/Http/Controllers/ManagerTimeClockDeviceController.php
@@ -9,6 +9,7 @@ use App\Http\Requests\ManagerTimeClockDevice\UpdateRequest;
 use App\Http\Resources\ManagerTimeClockDeviceResource;
 use App\Http\Resources\ManagerTimeClockDeviceResourceCollection;
 use App\Repositories\Contracts\ManagerTimeClockDeviceRepository;
+use Illuminate\Auth\Access\AuthorizationException;
 
 class ManagerTimeClockDeviceController extends Controller
 {
@@ -31,9 +32,12 @@ class ManagerTimeClockDeviceController extends Controller
      *
      * @param IndexRequest $request
      * @return ManagerTimeClockDeviceResourceCollection
+     * @throws AuthorizationException
      */
     public function index(IndexRequest $request)
     {
+        $this->authorize('list', ManagerTimeClockDevice::class);
+
         $managerTimeClockDevices = $this->managerTimeClockDeviceRepository->findBy($request->all());
 
         return new ManagerTimeClockDeviceResourceCollection($managerTimeClockDevices);
@@ -44,9 +48,12 @@ class ManagerTimeClockDeviceController extends Controller
      *
      * @param  \Illuminate\Http\Request  $request
      * @return ManagerTimeClockDeviceResource
+     * @throws AuthorizationException
      */
     public function store(StoreRequest $request)
     {
+        $this->authorize('store', ManagerTimeClockDevice::class);
+
         $managerTimeClockDevice = $this->managerTimeClockDeviceRepository->save($request->all());
 
         return new ManagerTimeClockDeviceResource($managerTimeClockDevice);
@@ -57,9 +64,12 @@ class ManagerTimeClockDeviceController extends Controller
      *
      * @param ManagerTimeClockDevice $managerTimeClockDevice
      * @return ManagerTimeClockDeviceResource
+     * @throws AuthorizationException
      */
     public function show(ManagerTimeClockDevice $managerTimeClockDevice)
     {
+        $this->authorize('show', $managerTimeClockDevice);
+
         return new ManagerTimeClockDeviceResource($managerTimeClockDevice);
     }
 
@@ -69,9 +79,12 @@ class ManagerTimeClockDeviceController extends Controller
      * @param UpdateRequest $request
      * @param ManagerTimeClockDevice $managerTimeClockDevice
      * @return ManagerTimeClockDeviceResource
+     * @throws AuthorizationException
      */
     public function update(UpdateRequest $request, ManagerTimeClockDevice $managerTimeClockDevice)
     {
+        $this->authorize('update', $managerTimeClockDevice);
+
         $managerTimeClockDevice = $this->managerTimeClockDeviceRepository->update($managerTimeClockDevice, $request->all());
 
         return new ManagerTimeClockDeviceResource($managerTimeClockDevice);
@@ -82,9 +95,12 @@ class ManagerTimeClockDeviceController extends Controller
      *
      * @param ManagerTimeClockDevice $managerTimeClockDevice
      * @return \Illuminate\Http\Response
+     * @throws AuthorizationException
      */
     public function destroy(ManagerTimeClockDevice $managerTimeClockDevice)
     {
+        $this->authorize('destroy', $managerTimeClockDevice);
+
         $this->managerTimeClockDeviceRepository->delete($managerTimeClockDevice);
 
         return response()->json(null, 204);
